src/research/competitor_analyzer.py
"""
Competitor Analyzer Module
==========================
Analyze competitor channels for insights
"""

from typing import Dict, List, Optional
from dataclasses import dataclass


@dataclass
class ChannelAnalysis:
    """채널 분석 데이터"""
    channel_name: str
    subscribers: int
    avg_views: int
    upload_frequency: str
    top_topics: List[str]
    content_style: str
    strengths: List[str]
    weaknesses: List[str]


@dataclass
class ContentGap:
    """콘텐츠 갭 분석"""
    topic: str
    opportunity_score: float
    competitor_coverage: str
    recommended_angle: str


class CompetitorAnalyzer:
    """경쟁 채널 분석기"""

    def __init__(self, config: Dict):
        self.config = config
        self.youtube_api = None
        self._init_api()

    def _init_api(self):
        """YouTube API 초기화"""
        try:
            from googleapiclient.discovery import build
            import os
            api_key = os.getenv('YOUTUBE_API_KEY')
            if api_key:
                self.youtube_api = build('youtube', 'v3', developerKey=api_key)
        except Exception:
            pass

    async def analyze_channel(self, channel_id: str) -> ChannelAnalysis:
        """
        채널 분석

        Args:
            channel_id: YouTube 채널 ID

        Returns:
            채널 분석 데이터
        """
        # 실제로는 YouTube API 사용
        return ChannelAnalysis(
            channel_name="Sample Channel",
            subscribers=100000,
            avg_views=50000,
            upload_frequency="weekly",
            top_topics=["history", "science"],
            content_style="educational",
            strengths=["High quality visuals", "Good narration"],
            weaknesses=["Irregular upload schedule"]
        )

    async def find_content_gaps(
        self,
        topic: str,
        competitor_channels: List[str]
    ) -> List[ContentGap]:
        """
        콘텐츠 갭 찾기

        Args:
            topic: 분석할 주제
            competitor_channels: 경쟁 채널 ID 리스트

        Returns:
            콘텐츠 갭 리스트
        """
        return [
            ContentGap(
                topic=f"{topic} - 미다뤄진 측면",
                opportunity_score=0.8,
                competitor_coverage="low",
                recommended_angle="심층 분석"
            )
        ]

    async def benchmark_performance(
        self,
        our_channel_id: str,
        competitor_ids: List[str]
    ) -> Dict:
        """성과 벤치마킹"""
        return {
            "our_avg_views": 10000,
            "competitor_avg_views": 50000,
            "gap_analysis": "Need to improve thumbnail CTR"
        }

    async def get_trending_in_niche(self, category: str) -> List[Dict]:
        """니치 내 트렌딩 콘텐츠"""
        return []
